requirements.txt
bpython
django
django-picklefield
docutils
